==> procedures/2-clean-remaining-data.ipynb <==
from helpers.dataframe_cleaner import *
from dotenv import load_dotenv
import os
load_dotenv()

path = os.getenv("KAGGLE_PATH")

fishs = pd.read_csv(f"{path}/dataset-ikan.csv")
lambs = pd.read_csv(f"{path}/dataset-kambing.csv")
tofus = pd.read_csv(f"{path}/dataset-tahu.csv")
eggs = pd.read_csv(f"{path}/dataset-telur.csv")
soybeans = pd.read_csv(f"{path}/dataset-tempe.csv")
prawns = pd.read_csv(f"{path}/dataset-udang.csv")

dataframes = [fishs, lambs, tofus, eggs, soybeans, prawns]
filenames = ["fishs_cleaned", "lambs_cleaned", "tofus_cleaned", "eggs_cleaned", "soybeans_cleaned", "prawns_cleaned"]
for i in range(dataframes.__len__()):
    df_clean_missing = clean_missing(dataframes[i])
    df_cleaned = clean_title_dupes(df_clean_missing)
    df_cleaned.to_csv(f"../csv/cleaned/{filenames[i]}.csv", index=False)
